requirements.txt
requests
jieba
xlrd
xlwt
baidu-aip
pyecharts==1.5.1